7_final_task/requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
